// src/servers/pexel/pexel-server-enums.ts
export enum getPexelPhotoByIdEnum {
  FOUND_PHOTOS = "FOUND_PHOTOS",
  NOT_PHOTO_TYPE = "NOT_PHOTO_TYPE",
  NULL = "NULL",
}

export enum getPexelVideoByIdEnum {
  FOUND_VIDEOS = "FOUND_VIDEOS",
  NOT_VIDEO_TYPE = "NOT_VIDEO_TYPE",
  NULL = "NULL",
}

export enum GetVideosEnum {
  VIDEOS_FOUND = "VIDEOS_FOUND",
  FAILED_TO_GET_ALL_VIDEOS = "FAILED_TO_GET_ALL_VIDEOS",
}

export enum GetVideosByKeywordEnum {
  VIDEOS_FOUND = "VIDEOS_FOUND",
  FAILED_TO_GET_VIDEOS = "FAILED_TO_GET_VIDEOS",
}